+++ ceammc/ext/doc/docs.cmake
@@ -536,6 +536,7 @@ set(DOC_FILES
     synth.dx7~
     synth.eguitar~
     synth.fgrain~
+    synth.glitch~
     synth.ks~
     synth.marimba~
     synth.rhodey~