+++ ceammc/ext/src/synth/mod_synth.cpp
@@ -12,6 +12,7 @@ extern void setup_synth_church_bell_tilde();
 extern void setup_synth_dubdub_tilde();
 extern void setup_synth_dx7_tilde();
 extern void setup_synth_eguitar_tilde();
+extern void setup_synth_glitch();
 extern void setup_synth_ks_tilde();
 extern void setup_synth_marimba_tilde();
 extern void setup_synth_risset_arp_tilde();
@@ -20,6 +21,7 @@ extern void setup_synth_russian_bell_tilde();
 
 void ceammc_synth_setup()
 {
+    // STK synths
     setup_synth_bthree();
     setup_synth_fgrain();
     setup_synth_rhodey();
@@ -27,6 +29,7 @@ void ceammc_synth_setup()
     setup_synth_sitar();
     setup_synth_wurley();
 
+    // faust synth
     setup_synth_birds_tilde();
     setup_synth_church_bell_tilde();
     setup_synth_dubdub_tilde();
@@ -37,4 +40,7 @@ void ceammc_synth_setup()
     setup_synth_risset_arp_tilde();
     setup_synth_risset_tone_tilde();
     setup_synth_russian_bell_tilde();
+
+    // Glitch
+    setup_synth_glitch();
 }