+++ ceammc/gui/plugins/ceammc_popup_menu_tree.tcl
@@ -219,8 +219,8 @@
     }}
     {synth {
         synth.bee3~ synth.birds~ synth.church_bell~ synth.dubdub~ synth.dx7~
-        synth.eguitar~ synth.fgrain~ synth.ks~ synth.marimba~ synth.rhodey~
-        synth.risset_arp~ synth.risset_tone~ synth.russian_bell~
+        synth.eguitar~ synth.fgrain~ synth.glitch~ synth.ks~ synth.marimba~
+        synth.rhodey~ synth.risset_arp~ synth.risset_tone~ synth.russian_bell~
         synth.shakers~ synth.sitar~ synth.wurley~
     }}
     {system {