+++ ceammc/CHANGELOG.md
@@ -67,6 +67,7 @@
     - synth.wurley~ ...
     - synth.sitar~ added
     - synth.russian_bell~ added
+    - synth.glitch~ added
     
 
 ### Changed: